third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_lineboundary_18.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'lineboundary'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} lineboundary`
  );
}

function makeSample(offset){
  const kSample = 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right lineboundary
testIt('ltr', 0, 'right', '^they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 1, 'right', 't^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 2, 'right', 'th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 3, 'right', 'the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 4, 'right', 'they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 5, 'right', 'they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 6, 'right', 'they s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 7, 'right', 'they sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 8, 'right', 'they sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 9, 'right', 'they said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 10, 'right', 'they said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 11, 'right', 'they said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 12, 'right', 'they said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 13, 'right', 'they said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 14, 'right', 'they said "\u202B;c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 15, 'right', 'they said "\u202B;ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 16, 'right', 'they said "\u202B;car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 17, 'right', 'they said "\u202B;car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 18, 'right', 'they said "\u202B;car \u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 19, 'right', 'they said "\u202B;car \u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 20, 'right', 'they said "\u202B;car \u05D3;\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 21, 'right', 'they said "\u202B;car \u05D3;\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 22, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 23, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 24, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 25, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 26, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 27, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2;\u202C;."|');
testIt('ltr', 28, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2;\u202C;."|');
testIt('ltr', 29, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;."|');
testIt('ltr', 30, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;."|');
testIt('ltr', 31, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;."|');
testIt('ltr', 32, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^."|');
testIt('ltr', 33, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;.^"|');

// ltr left lineboundary
testIt('ltr', 0, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 1, 'left', '|t^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 2, 'left', '|th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 3, 'left', '|the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 4, 'left', '|they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 5, 'left', '|they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 6, 'left', '|they s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 7, 'left', '|they sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 8, 'left', '|they sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 9, 'left', '|they said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 10, 'left', '|they said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 11, 'left', '|they said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 12, 'left', '|they said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 13, 'left', '|they said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 14, 'left', '|they said "\u202B;c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 15, 'left', '|they said "\u202B;ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 16, 'left', '|they said "\u202B;car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 17, 'left', '|they said "\u202B;car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 18, 'left', '|they said "\u202B;car \u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 19, 'left', '|they said "\u202B;car \u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 20, 'left', '|they said "\u202B;car \u05D3;\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 21, 'left', '|they said "\u202B;car \u05D3;\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 22, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 23, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 24, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 25, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 26, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1;\u05D2;\u202C;."');
testIt('ltr', 27, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2;\u202C;."');
testIt('ltr', 28, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2;\u202C;."');
testIt('ltr', 29, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;."');
testIt('ltr', 30, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;."');
testIt('ltr', 31, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;."');
testIt('ltr', 32, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^."');
testIt('ltr', 33, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;.^"');

// ltr forward lineboundary
testIt('ltr', 0, 'forward', '^they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 1, 'forward', 't^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 2, 'forward', 'th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 3, 'forward', 'the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 4, 'forward', 'they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 5, 'forward', 'they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 6, 'forward', 'they s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 7, 'forward', 'they sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 8, 'forward', 'they sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 9, 'forward', 'they said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 10, 'forward', 'they said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 11, 'forward', 'they said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 12, 'forward', 'they said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 13, 'forward', 'they said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 14, 'forward', 'they said "\u202B;c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 15, 'forward', 'they said "\u202B;ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 16, 'forward', 'they said "\u202B;car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 17, 'forward', 'they said "\u202B;car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 18, 'forward', 'they said "\u202B;car \u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 19, 'forward', 'they said "\u202B;car \u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 20, 'forward', 'they said "\u202B;car \u05D3;\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 21, 'forward', 'they said "\u202B;car \u05D3;\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 22, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 23, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 24, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 25, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 26, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1;\u05D2;\u202C;."|');
testIt('ltr', 27, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2;\u202C;."|');
testIt('ltr', 28, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2;\u202C;."|');
testIt('ltr', 29, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;."|');
testIt('ltr', 30, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;."|');
testIt('ltr', 31, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;."|');
testIt('ltr', 32, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^."|');
testIt('ltr', 33, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;.^"|');

// ltr backward lineboundary
testIt('ltr', 0, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 1, 'backward', '|t^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 2, 'backward', '|th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 3, 'backward', '|the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 4, 'backward', '|they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 5, 'backward', '|they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 6, 'backward', '|they s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 7, 'backward', '|they sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 8, 'backward', '|they sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 9, 'backward', '|they said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 10, 'backward', '|they said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 11, 'backward', '|they said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 12, 'backward', '|they said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 13, 'backward', '|they said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 14, 'backward', '|they said "\u202B;c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 15, 'backward', '|they said "\u202B;ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 16, 'backward', '|they said "\u202B;car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 17, 'backward', '|they said "\u202B;car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 18, 'backward', '|they said "\u202B;car \u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 19, 'backward', '|they said "\u202B;car \u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 20, 'backward', '|they said "\u202B;car \u05D3;\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 21, 'backward', '|they said "\u202B;car \u05D3;\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 22, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 23, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 24, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 25, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1;\u05D2;\u202C;."');
testIt('ltr', 26, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1;\u05D2;\u202C;."');
testIt('ltr', 27, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2;\u202C;."');
testIt('ltr', 28, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2;\u202C;."');
testIt('ltr', 29, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;."');
testIt('ltr', 30, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;."');
testIt('ltr', 31, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;."');
testIt('ltr', 32, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^."');
testIt('ltr', 33, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;.^"');

// rtl right lineboundary
testIt('rtl', 0, 'right', '^they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 1, 'right', 't^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 2, 'right', 'th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 3, 'right', 'the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 4, 'right', 'they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 5, 'right', 'they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 6, 'right', 'they s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 7, 'right', 'they sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 8, 'right', 'they sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 9, 'right', 'they said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 10, 'right', 'they said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 11, 'right', 'they said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 12, 'right', 'they said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 13, 'right', 'they said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 14, 'right', 'they said "\u202B;c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 15, 'right', 'they said "\u202B;ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 16, 'right', 'they said "\u202B;car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 17, 'right', 'they said "\u202B;car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 18, 'right', 'they said "\u202B;car \u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 19, 'right', 'they said "\u202B;car \u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 20, 'right', 'they said "\u202B;car \u05D3;\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 21, 'right', 'they said "\u202B;car \u05D3;\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 22, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 23, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 24, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 25, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 26, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 27, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2;\u202C;."|');
testIt('rtl', 28, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2;\u202C;."|');
testIt('rtl', 29, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;."|');
testIt('rtl', 30, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;."|');
testIt('rtl', 31, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;."|');
testIt('rtl', 32, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^."|');
testIt('rtl', 33, 'right', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;.^"|');

// rtl left lineboundary
testIt('rtl', 0, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 1, 'left', '|t^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 2, 'left', '|th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 3, 'left', '|the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 4, 'left', '|they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 5, 'left', '|they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 6, 'left', '|they s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 7, 'left', '|they sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 8, 'left', '|they sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 9, 'left', '|they said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 10, 'left', '|they said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 11, 'left', '|they said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 12, 'left', '|they said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 13, 'left', '|they said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 14, 'left', '|they said "\u202B;c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 15, 'left', '|they said "\u202B;ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 16, 'left', '|they said "\u202B;car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 17, 'left', '|they said "\u202B;car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 18, 'left', '|they said "\u202B;car \u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 19, 'left', '|they said "\u202B;car \u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 20, 'left', '|they said "\u202B;car \u05D3;\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 21, 'left', '|they said "\u202B;car \u05D3;\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 22, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 23, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 24, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 25, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 26, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1;\u05D2;\u202C;."');
testIt('rtl', 27, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2;\u202C;."');
testIt('rtl', 28, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2;\u202C;."');
testIt('rtl', 29, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;."');
testIt('rtl', 30, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;."');
testIt('rtl', 31, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;."');
testIt('rtl', 32, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^."');
testIt('rtl', 33, 'left', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;.^"');

// rtl forward lineboundary
testIt('rtl', 0, 'forward', '^they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 1, 'forward', 't^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 2, 'forward', 'th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 3, 'forward', 'the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 4, 'forward', 'they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 5, 'forward', 'they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 6, 'forward', 'they s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 7, 'forward', 'they sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 8, 'forward', 'they sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 9, 'forward', 'they said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 10, 'forward', 'they said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 11, 'forward', 'they said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 12, 'forward', 'they said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 13, 'forward', 'they said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 14, 'forward', 'they said "\u202B;c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 15, 'forward', 'they said "\u202B;ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 16, 'forward', 'they said "\u202B;car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 17, 'forward', 'they said "\u202B;car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 18, 'forward', 'they said "\u202B;car \u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 19, 'forward', 'they said "\u202B;car \u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 20, 'forward', 'they said "\u202B;car \u05D3;\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 21, 'forward', 'they said "\u202B;car \u05D3;\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 22, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 23, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 24, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 25, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 26, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1;\u05D2;\u202C;."|');
testIt('rtl', 27, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2;\u202C;."|');
testIt('rtl', 28, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2;\u202C;."|');
testIt('rtl', 29, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;."|');
testIt('rtl', 30, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;."|');
testIt('rtl', 31, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;."|');
testIt('rtl', 32, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^."|');
testIt('rtl', 33, 'forward', 'they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;.^"|');

// rtl backward lineboundary
testIt('rtl', 0, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 1, 'backward', '|t^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 2, 'backward', '|th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 3, 'backward', '|the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 4, 'backward', '|they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 5, 'backward', '|they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 6, 'backward', '|they s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 7, 'backward', '|they sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 8, 'backward', '|they sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 9, 'backward', '|they said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 10, 'backward', '|they said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 11, 'backward', '|they said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 12, 'backward', '|they said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 13, 'backward', '|they said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 14, 'backward', '|they said "\u202B;c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 15, 'backward', '|they said "\u202B;ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 16, 'backward', '|they said "\u202B;car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 17, 'backward', '|they said "\u202B;car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 18, 'backward', '|they said "\u202B;car \u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 19, 'backward', '|they said "\u202B;car \u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 20, 'backward', '|they said "\u202B;car \u05D3;\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 21, 'backward', '|they said "\u202B;car \u05D3;\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 22, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 23, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 24, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 25, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1;\u05D2;\u202C;."');
testIt('rtl', 26, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1;\u05D2;\u202C;."');
testIt('rtl', 27, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2;\u202C;."');
testIt('rtl', 28, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2;\u202C;."');
testIt('rtl', 29, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;."');
testIt('rtl', 30, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;."');
testIt('rtl', 31, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;."');
testIt('rtl', 32, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^."');
testIt('rtl', 33, 'backward', '|they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;.^"');
</script>
